End of section 10

+++ Application/Activities/Delete.cs
@@ -1,4 +1,5 @@
 
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,12 +7,12 @@ namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
             public Guid Id { get; set; }
         }
 
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -20,11 +21,24 @@ namespace Application.Activities
                 _context = context;
             }
 
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
+
+                if (activity is null)
+                {
+                    return null;
+                }
+
                 _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync(cancellationToken);
+                var result = await _context.SaveChangesAsync(cancellationToken) > 0;
+
+                if (!result)
+                {
+                    return Result<Unit>.Failure("Failed to delete activity");
+                }
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
